move the relative game to a new games folder
